+++ ipc/package/CMakeLists.txt
@@ -1,9 +1,9 @@
 add_library(input_stream INTERFACE)
 target_include_directories(input_stream INTERFACE ${CMAKE_CURRENT_LIST_DIR}/include)
 
-add_library(ring_queue INTERFACE)
-target_include_directories(ring_queue INTERFACE ${CMAKE_CURRENT_LIST_DIR}/include)
-target_link_libraries(ring_queue INTERFACE input_stream)
+add_library(ring_buffer_input_stream INTERFACE)
+target_include_directories(ring_buffer_input_stream INTERFACE ${CMAKE_CURRENT_LIST_DIR}/include)
+target_link_libraries(ring_buffer_input_stream INTERFACE input_stream)
 
 add_library(package_utils INTERFACE)
 target_sources(package_utils INTERFACE ${CMAKE_CURRENT_LIST_DIR}/src/package_utils.cpp)
