+++ ipc/package/include/ring_buffer_input_stream.hpp
