+++ ipc/package/tests/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_executable(
 	package_tests
-	${CMAKE_CURRENT_LIST_DIR}/src/ut_ring_queue.cpp
+	${CMAKE_CURRENT_LIST_DIR}/src/ut_ring_buffer_input_stream.cpp
 	${CMAKE_CURRENT_LIST_DIR}/src/ut_package_reader.cpp
 	${CMAKE_CURRENT_LIST_DIR}/src/ut_package_writer.cpp
 )
@@ -8,7 +8,7 @@ add_executable(
 target_link_libraries(
 	package_tests
 	PRIVATE
-	ring_queue
+	ring_buffer_input_stream
 	package_reader
 	package_writer
 	package_utils
